Refactor of vllm_ascend/distributed module
